requirements.txt
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
statsmodels==0.14.1
openpyxl==3.1.2
streamlit==1.32.2
python-dotenv==1.0.1
